Start work on amd64 port
- Create x86_64 arch and pc64 platform stubs (copied from the 32-bit
  equivalents; this doesn't even link since a lot of stuff isn't
  implemented, but the kernel code at least compiles as a whole)

+++ kernel/src/sched/Scheduler.cpp
@@ -284,7 +284,7 @@ void Scheduler::receivedDispatchIpi(const uintptr_t irqToken) {
 
     // define the callback for irq acknowledgement
     auto ackFxn = [](void *ctx) {
-        const auto irq = reinterpret_cast<uint32_t>(ctx);
+        const auto irq = reinterpret_cast<uintptr_t>(ctx);
         platform_irq_ack(irq);
     };
